slideshow: fix para. rendering when using SdrBlockTextPrimitive2D
If using an empty ViewInformation, it is possible that the prim.
decomposition is redone, when a different ViewInformation is used.
This removes all visibility flags for the paragraphs, so the whole
text is rendered. This fixes the issue by using the same instance
of the ViewInformation that is also used later for rendering the
primitives to the VirtualDevice, so the buffered decomposition is
used.

Also a test is added - it checks the hash of the buffer, which
should be different for all 3 layers. For this the tostring has
to be moved to the common hash.hxx, so we can reuse it in out
tests.

Change-Id: I4c951215a52e302d3b7b60a30c1b995002e53a4b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174989
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Szymon Kłos <szymon.klos@collabora.com>