Drop extra quoting
...introduced in 074714fab8db751e2aa0138391a270124733977d ".NET Bindings: Create
nuget package for LO SDK".  The assumption is that none of the variables used in
that block of code represent paths that contain characters that would need any
form of quoting.  (And if they did, the solution would not be to just wrap them
in "..." quoting, but to use '...' and escape any contained ' properly.)

Change-Id: Ie5b972c53a60c8c9a4bcfa03627a8e1998646285
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/175010
Reviewed-by: Stephan Bergmann <stephan.bergmann@allotropia.de>
Tested-by: Jenkins