tdf#163486: PVS: redundant nullptr check
V668 	There is no sense in testing the 'indxc' pointer against null, as the memory was allocated using the 'new' operator. The exception will be generated in the case of memory allocation error.
V668 	There is no sense in testing the 'indxr' pointer against null, as the memory was allocated using the 'new' operator. The exception will be generated in the case of memory allocation error.
V668 	There is no sense in testing the 'ipiv' pointer against null, as the memory was allocated using the 'new' operator. The exception will be generated in the case of memory allocation error.

Change-Id: Ifecd9ef5b831e5b312bb428e4d4f72c03cc29b7b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/175106
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>
Tested-by: Jenkins