tdf#163486: PVS: redundant nullptr check
V668 	There is no sense in testing the 'mpMarkedRanges' pointer against null, as the memory was allocated using the 'new' operator. The exception will be generated in the case of memory allocation error.
V668 	There is no sense in testing the 'mpMarkedRanges' pointer against null, as the memory was allocated using the 'new' operator. The exception will be generated in the case of memory allocation error.

Change-Id: Iba55f15534cdc9af5b86b288b5c42702ad83dc26
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/175107
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>
Tested-by: Jenkins