tdf#130857 qt weld: Implement QtInstanceDialog::response
Move the QtInstanceMessageDialog::response
implementation down into the base class
QtInstanceDialog and simplify a bit, and drop
the override in QtInstanceMessageDialog.

This will be needed e.g. by the
svtools/uiconfig/ui/restartdialog.ui
dialog shown when toggling experimental mode in
"Tools" -> "Options" and closing the dialog with
the "OK" button.

Change-Id: I147ea68554725a9d3b6bea6f77b15e7754cc962b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/175132
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>