Windows/--disable-skia: fix D2DWriteTextOutRenderer to properly set AA mode
This simplifies the process; in D2DWriteTextOutRenderer ctor, it
uses the final mode for the call to CreateRenderTarget, applying
the correct mode.

Also, pass the AA flag from SalGraphics, and check the setting of
GetUseFontAAFromSystem to use it, as done in e.g. CairoTextRender
after commit e6538f5bdd876911ea30f84a6512c03908e620fd (tdf#118966
vcl: add a flag to determine if AA of fonts is used from the system,
2018-07-28).

This fixes the failures on Windows with --disable-skia, as seen in
https://lists.freedesktop.org/archives/libreoffice/2024-October/092541.html

VclCjkTextTest::testVerticalText needed to be fixed by a tweak in
getCharacterRightSideHeight, to use color's IsDark, instead of
comparing to the fixed black, because the correct AA mode that is
used now, makes the vertical part of the character not completely
black.

Change-Id: Iee8fe98e29a80a242f8e761c9a23c68b34a45699
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/175188
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
(cherry picked from commit 85a56048a004b85d232805cc3b15db0c100d8a5c)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/175235
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>