tdf#151941: Show Alt_Shift_P as shortcut key for Page Style
In Writer the Tools>Customize>Keyboard list did not show attribution
for Alt_Shift_P which is the binding for Page Style dialog. The reason
for this was that .xcu node for the shortcut was part of Global node.

Change the Accelerators.xcu file to Modules node particularly,
under TextDocument com.sun.star.text.TextDocument.

Change-Id: I997c4e6dd292c46fe067a59f0309681d6c346a6b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/175249
Tested-by: Jenkins
Reviewed-by: Heiko Tietze <heiko.tietze@documentfoundation.org>