improve style searching in SvXMLStylesContext
We can searching without needing a std::map by sorting
the style list. Which also allows to do prefix
searching. Which we can use to dramatically reduce
the number of styles we need to loop through in
SdXMLStylesContext::ImpSetGraphicStyles.

a similar mega-master-page scenario as reported in tdf#158773

13.5 - 9.1s

Needed to adjust some unit tests because
the order of iteration through styles is
now different, which affects some file output.

Change-Id: Ia7240fe520b70839d2519eba1fb70819a3c3bf81
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/175281
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Tested-by: Jenkins