tdf#152534 Win 11 Dark: Draw toolbar button bg + frame manually
When using the dark theme on Windows 11, the background
color for toolbar buttons is light blue as with the light
theme, which results in insufficient contrast.

According to the comments in tdf#152534, this is
is a in the Windows dark theme which can also be seen in
other apps or has been worked around there, like in
Notepad++ [1].

Work around the issue by not letting the theme draw the
background (using Win API function `DrawThemeBackground`)
when on Windows 11 and using dark theme, but manually draw
the background and a frame instead for now.

For the background color, use a color that's a bit
lighter than the background color for non-highlighted
items.

[1] https://github.com/notepad-plus-plus/notepad-plus-plus/commit/5d086f93a80d275fcc5e4047ae49c614680a7e5b

Change-Id: Ie35937fde2e8c7078c4979a2ef60c28fc4679574
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/175334
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins