tdf#130857 qt weld: Create QGridLayout for "GtkGrid"
Create a QGridLayout [1] when encountering a "GtkGrid"
object in a .ui file.

This will be needed e.g. by Writer's "Tools" -> "Word Count"
dialog.

With this commit in place, adding "modules/swriter/ui/wordcount.ui"
to the list of supported .ui files in
QtInstanceBuilder::IsUIFileSupported would already result in each of
the labels in the dialog showing up, but each one as a single row rather
than they being properly arranged in rows and columns.
(That will be handled in an upcoming commit.)

[1] https://doc.qt.io/qt-6/qgridlayout.html

Change-Id: Ib9a9de4aa2820ac7e6771acf884072768508fe59
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/175365
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>