tdf#130857 qt weld: Evaluate grid positions
Implement handling for the "left-attach" and
"top-attach" packing properties of "GtkGrid"
children, which describe the column and row
of these children within the grid.

Introduce a new static helper method
QtBuilder::applyGridPackingProperties that
implements the handling for QWidget children
for now.

(Support for QLayout items within a grid
will have to be added later in order to
supported cases where e.g. a "GtkBox" is
located inside of a "GtkGrid" in a .ui file.)

In order to move the item to the proper position
within the grid, first locate it to determine
it's current row and column index, remove the item
from the layout, and re-insert it at the new
position.

While this might not be the most efficient
way of doing this, it is fairly easy to
implement, without having to change the
overall approach that QtBuilder, VclBuilder and the
WidgetBuilder base currently process .ui
files.

This is sufficient for Writer's "Word Count"
dialog, for which support will be declared
in an upcoming commit.

Change-Id: Ia296373c408e6cd84ffcc29b9d9a03d3c2441816
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/175366
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins