cool#9992 lok doc sign: allow late-init of the sign cert
The current setup is that doc_initializeForRendering() gets the signing
key/cert earlier, even if no signing will be performed for this
document.

This has the downside that we needlessly share sensitive info.

Add an alternative way so to only share the sign key/cert when the
.uno:Signature command gets dispatched. This works similarly well for
the signing, but this way the private info is only shared when we
actually use it.

The .uno:Signature UNO command brings up the interactive dialog on
success, so it's not easy to cover this from CppunitTest_sfx2_doc.
Anyhow, the format for these two parameters is PEM, i.e. base64 dump of
the data, with the usual markers to show which one is the key and the
cert.

Change-Id: Ib424a1a490a3eb4aab35bc70a50791fc0d400920
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/175373
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins