diff-pdf-page.py: change assumed background to transparent
The page background color of a document likely is white,
so at least the overlay (the second PNG) must be transparent.

[I saw (in Draw) that PRINT-TO-PDF tended to produce a PDF
that would have a transparent background when converted to PNG,
but that EXPORT-TO-PDF from LO would produce a white background
that obviously would fully overlay and obscure the first image.]

(Calc and Writer's export-to-pdf produced a transparent background.)

In any case, it makes sense to try to force both of these
to have a transparent background.

NOTE: It was important to place -transparent before -fuzz,
or else 95% of all colors will be considered white
and turned into transparent.

Change-Id: I26e6081fcd59e82b2aaedb86370bdbcc57b0a24e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/175466
Reviewed-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins
Reviewed-by: László Németh <nemeth@numbertext.org>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/175563
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>