Try to apply indents a bit saner
It was too convoluted; the three properties that were set there
all used different logic (PROP_PARA_FIRST_LINE_INDENT was only
set when the parent style id was present; PROP_PARA_LEFT_MARGIN
was set when it was found either in the entry or in its parent;
PROP_PARA_RIGHT_MARGIN was only set when parent style was found,
and it had properties). It seems to be just some leftovers from
iterative partial changes of the code.

This makes the code more uniform. For PROP_PARA_RIGHT_MARGIN it
keeps setting PROP_PARA_FIRST_LINE_INDENT/PROP_PARA_LEFT_MARGIN
from getNumberingProperty; but otherwise, all three are handled
similarly: if the property is set either in the entry itself or
in its parent, it is processed. This changes the logic slightly,
but hopefully in the correct direction.

Change-Id: I1a8fb95d055101dcbc3aaec6207721146c94cc21
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/175437
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Jenkins