slideshow: simplify rendering, render text fields each in own layer
This change significantly simplifies the rendering of layers by
introducing an analysis pass (when we render the background) that
creates rendering pass objects, which contain which objects and
paragraphs need to be rendered in each pass. Previously we tried
to figure this out durign the rendering, which meant we had to
track a lot of various states. This all is not longer needed. This
chaneg also means that the new RenderPassObjectRedirector is fairly
simple.

In addition this changes the test and implementation to render
each text field in its own layer as it is expected by the online
implementation (to enable master pages caching), but it still
doesn't first render the placeholders for text fields as MasterPage
layer and the TextField layer as a separate bitmap layer. This
will be done next.

Change-Id: I5377f4c5e5ebff67da70b248c831837cb9f2e559
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/175568
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>