win a11y: No longer read/write registry key
No longer attempt to write a registry
"SupportAssistiveTechnology" in
"HKEY_CURRENT_USER\Software\LibreOffice\Accessibility\AtToolSupport"
when assistive technology support gets enabled (e.g. because a
screen reader is running) and no longer try to read that registry
key to determine whether or not AT tool support is requested.

While the comment

    // Check in the Windows registry if an AT tool wants Accessibility support to
    // be activated ..

seems to suggest that this is something requested by an assistive
technology, it looks more like a way to persist whether the a11y
bridge was ever used in LO, as LO itself writes that value.

I don't really see the point, but think it makes much more
sense to enable the a11y bridge if an AT is active
*at the moment*.

In addition, in my Windows 10 setup with LO 24.8 installed, there's no
"HKEY_CURRENT_USER\Software\LibreOffice" in the registry, only a
"HKEY_CURRENT_USER\Software\The Document Foundation\LibreOffice 24.8",
so the code wouldn't have any effect in practice anyway.

Maybe this was necessary at some point in the past
when the Java Access Bridge was still in place, instead
of supporting the platform API via today's MSAA/IAccessible2
bridge

Note that before

    commit bfbaeb8192447265bdd78d1be4990947d135eb6e
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Fri Oct 18 17:01:43 2024 +0100

        tdf#160982 wina11y: Drop extra screen reader check to enable a11y bridge

, the presence or absence of the SPI_GETSCREENREADER param
or a SAL_FORCE_IACCESSIBLE2 environment variable would
additionally have played a role.

Change-Id: I4dddb599eedb4b29c9709fbb41093ef615e60b8f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/175501
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>