sw a11y: Don't prevent zooming into print preview when AT active
In Writer's print preview ("File" -> "Print Preview"),
allow zooming in or out using Ctrl + mouse wheel
regardless of whether assistive technology is enabled
(on Windows) or not.

I see no reason to prevent this when AT is active.

This was originally introduced in

    commit 70de0c38fd177b3c95837118c5fb35e380c87e90
    Author: Jens-Heiner Rechtien <hr@openoffice.org>
    Date:   Thu Mar 27 14:45:43 2003 +0000

        MWS_SRX644: migrate branch mws_srx644 -> HEAD

with a

    //#106746# zoom has to be disabled if Accessibility support is switched on

comment, but that unfortunately only refers to a ticket
in the StarDivision internal issue tracker, so the
reasoning from back then remains unclear.

(`git show --ignore-space-change` to see the "actual
change" more clearly)

Change-Id: Iccec17a334dbe19de14cf7889d7f70f0051f152e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/175502
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins