tdf#130857 qt weld: Signal when current combobox item changes
Connect to the QComboBox::currentIndexChanged signal and
call `weld::ComboBox::signal_changed` to notify of that change.

This is in preparation for upcoming

    Change-Id: I7806688102f690faa02fb5e712943d6ae216ff9a
    Author: Michael Weghorn <m.weghorn@posteo.de>

        tdf#130857 qt weld: Declare support for "Printer Settings" dialog

and causes the infos for the printer previously selected
in the combobox in the "File" -> "Printer Settings" dialog
to be shown after selecting a printer other than the initially
selected one in the combobox, then clicking the "Properties"
button in that dialog.
(Without this, properties of the initially selected printer would
always be shown.)

Change-Id: I9a8609e172fd1d6783a86395ff58e64b7cf5c4d2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/175521
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>