tdf#130857 qt: Move pixmap helper to new include/vcl/qt/QtUtils.hxx
Introduce include/vcl/qt/QtUtils.hxx as a new header
for Qt-specific utility functions that can be used
in multiple modules, e.g. avmedia and the Qt VCL
plugins in vcl.

This is meant to be used only by code that already
links in Qt.

There's already vcl/inc/qt{5,6}/QtTools.hxx for
helpers needed in vcl only.

Initially, add a `loadQPixmapIcon` helper function
that can be used to retrieve a QPixmap for an icon name
and is extracted from QtPlayer::createMediaPlayerWidget
in avmedia.

It will be reused to implement the
QtInstanceButton::set_from_icon_name logic in an upcoming
commit.

Change-Id: I9f25aa5ca8f00da97d06ecdd164a8fae10e492dd
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/175524
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>