tdf#130857 qt weld: Handle entry for editable comboboxes
An editable GtkComboBox has a "has-entry" property of "true"
and an internal GtkEntry child.

Make the QComboBox editable depending on that "has-entry" property.
Mark the QLineEdit object created for the "GtkEntry" for deletion
again, as the QComboBox already has a QLineEdit by itself if it's
editable, and the one created for the internal child would otherwise
be useless and oddly overlap the combobox.

(Seen e.g. with the "Tools" -> "Options" -> "Languages and Locales"
-> "Writing Aids" -> "New" dialog" in a WIP branch where that dialog's
.ui file ("cui/ui/optnewdictionarydialog.ui") was added to the list of
supported files for QtInstanceBuilder. More work is needed
on other aspects still.)

Change-Id: I25ea74c732e60f50035604a4fc75dad50f4cf55f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/175531
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>