tdf#153182 PPTX Chart with datamodel rendered incorrectly
Partial fix.

(*) fixed bugs where the use of continue in loops meant that a tracking
variable was not being updated

(*) improved exception message in TableModel::getCellByPosition

(*) rather than exiting the view generation process with an exception,
log a warning and skip that part of the view process

The root of the problem here is that we don't seem to be importing
data-tables from PPTX properly. I suspect that we need more import
code, in particular, we probably need to read some flag that
indicates if we have header rows and columns.

Change-Id: Iaa9a1593bc185a5bb4f36a434117bbd5bf323e18
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/175551
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>
Tested-by: Jenkins