tdf#130857 optnewdictionarydialog.ui: Define focusable widgets in order
Switch the order that the widgets are defined in the .ui file
to match their logical order (to be top to bottom, left to right)
as also described by their "left-attach" and "top-attach"
properties.

This only rearranges the order without changing anything else
for the single children and doesn't have any effect for the gtk3
and VCL implementations.

But as described in more details in

    commit 0fe7b3cebc1a070b13b825635f4bd387a22e41ab
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Wed Oct 23 22:43:10 2024 +0200

        tdf#130857 printersetupdialog.ui: Define focusable widgets in order

, Qt uses the construction order of widgets to determine
the tab order as well, meaning that focus would unexpectedly jump from
the edit to the checkbox, then back to the combobox without
this commit in place for the qt5 and qt6 VCL plugins in this
dialog ("Tools" -> "Options" -> "Languages and Locales" -> "Writing Aids"
-> "New") now that native Qt widgets are used.

Change-Id: Ib09137cb11dab60c4a1af4fbc3b007e801dea33b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/175567
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>