tdf#130857 qt weld: Prefer custom handler for dialog buttons
If a custom handler is set for a button in a dialog's
button box, skip the default handling (to close the
dialog with the corresponding response code, unless
it's the help button) in QtInstanceDialog::handleButtonClick,
so that the custom handler gets processed instead
(see QtInstanceButton::buttonClicked).

This is similar to what the VCL implementation
(see SalInstanceButton's ClickHdl) and the
gtk3 implementation (see
GtkInstanceDialog::asyncresponse) do.

In order to be able to determine whether
a custom click handler is set, make
weld::Button::connect_clicked virtual
and override it in QtInstanceButton::connect_clicked
to set a property on the QPushButton object
to indicate that a handler is set.
Introduce a static helper method
QtInstanceButton::hasCustomClickHandler
that evaluates this property.

Without this change in place, pressing the
"Rename" button in Writer's "Rename Object"
dialog that will be declared as supported
in upcoming commit

    Change-Id: I0ac8f66b977f5d999bb197cc68e8ce5533fe1ebd
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Fri Oct 25 10:41:45 2024 +0200

        tdf#130857 qt weld: Declare support for sw "Rename Object" dialog

would not actually result in a bookmark getting
renamed, only the dialog would be closed.

Change-Id: If3e762100349ecd06c4e9415e9a8f9a44ec6491c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/175633
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>