tdf#130857 qt weld: Add QtInstanceRadioButton
Implement initial support for native radio buttons
using QRadioButton:

* Let QtBuilder create a QRadioButton widget
  when it encounters a "GtkRadioButton" object.
* Let QtBuilder::setProperties also handle the
  QRadioButton case. Both, QRadioButton and
  QCheckBox derive from QAbstractButton, so reuse
  the existing logic for QCheckBox to set label and
  checked status.
* Add new class QtInstanceRadioButton as a weld::RadioButton
  implementation that uses a QRadioButton widget.
* Let QtInstanceBuilder::weld_radio_button return
  an instance of the new class.

For now, ignore the GtkRadioButton "group" property [1]
that is used to group radio buttons.

QRadioButton's are automatically grouped when they
have the same parent widget, which is sufficient
for the case of the "Alignment" dialog in Math
for which support will be declared in an upcoming
commit.

For more complex scenarios, the use of QButtonGroup [2]
could be implemented in the future to explicitly group
radio buttons, as mentioned in the QRadioButton doc [3]:

> If you need multiple exclusive button groups for radio buttons that
> belong to the same parent widget, put them into a QButtonGroup.

[1] https://docs.gtk.org/gtk3/property.RadioButton.group.html
[2] https://doc.qt.io/qt-6/qbuttongroup.html
[3] https://doc.qt.io/qt-6/qradiobutton.html

Change-Id: Iaf8b0fef00fc10268c09410080156e7913634ab7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/175639
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins