tdf#130857 qt weld: Add QtInstanceImage skeleton
Add a new class QtInstanceImage as a weld::Image
implementation using native Qt widgets.

It uses a QLabel that holds the Image, which can
be set using QLabel::setPixmap [1].

This commit adds the new class, lets
QtInstanceBuilder::weld_image return an instance
of the new class, and extends QtBuilder to create
a QLabel when it encounters a "GtkImage" object.

The logic to actually set the image when either
processing the .ui file or when one of the
weld::Image::set_from_icon_name or one of the
weld::Image::set_image variants gets called
still needs to be implemented in subsequent
commits for the label to actually display
an image.

[1] https://doc.qt.io/qt-6/qlabel.html#pixmap-prop

Change-Id: I3aa526b89487a5301dabd108723d7b68ca85ea66
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/175655
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins