tdf#130857 qt weld: Add a QtInstanceLinkButton
Add a new QtInstanceLinkButton class that is
the weld::LinkButton implementing using a native
Qt widget.

QLabel can be used to display a hyperlink, as it
supports the HTML syntax for the text, so using ,

    QLabel myLabel;
    myLabel.setText(QLatin1String("<a href=\"https://www.libreoffice.org">LibreOffice Website</a>"));

can be used to let the QLabel handle a hyperlink.

To make it simple to set this as needed for a QLabel,
implement a new QLabel subclass called QtHyperlinkLabel
that provides convenient getters and setters to set
the displayed text and the link target, and takes
care of setting the QLabel text based on that as needed.

Implement QtInstanceLinkButton using an instance
of that class as the widget and create an instance of
that class in QtBuilder when encountering a "GtkLinkButton"
object while processing a .ui file and evaluate
the "label" and "uri" properties.

Change-Id: I71d28b6e5e3cbd110ec5b3d1232d55e9d2bb8a1b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/175656
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins