tdf#130857 qt weld: Add child layouts to grid
If a QLayout object is created as the child of a
QGridLayout object (e.g. a "GtkBox" as the direct child
of a "GtkGrid" in a .ui file), add that layout to the
QGridLayout - initially in a new row, but rearranging
that will be done as part of processing the "packing"
properties of the children once that is implemented
for layouts as well. (See  QtBuilder::applyPackingProperties
and QtBuilder::applyGridPackingProperties where this
is currently only implemented for QWidget children).

This will be needed e.g. for the "Help" -> "About LibreOfficeDev"
dialog.

Change-Id: I292134b26efa0dfe5c8fe0684dc02dece338d579
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/175659
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>