tdf#130857 qt weld: Allow null func for async dialog
Drop the assert that a function to call is set
in QtInstanceDialog::dialogFinished and just skip
that step if none is set.

While `m_aRunAsyncFunc` gets set in both
QtInstanceDialog::runAsync variants and
QtInstanceDialog::dialogFinished shouldn't get
called without these having been called first,
QtInstanceDialog::runAsync can explicitly be called with
null for the function, which is the case e.g. for the
"Help" -> "About" dialog (s. the SID_ABOUT case in
SfxApplication::MiscExec_Impl), so that assert
triggers in a WIP branch for implementing support
for that dialog using native Qt widgets when the
dialog is closed.

Change-Id: I18578a4b1780b05f59817ad85eee3c3a20ba3b41
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/175660
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins