tdf#130857 qt weld: Simplify QtBuilder::applyGridPackingProperties
No longer try to retrieve the current row and column
of the widget in the QGridLayout and apply the "top-attach"
and "left-attach" properties to set the new row and column
separately.

Instead, assume (and assert), that both, the new row and column
properties are set, get them and then set the position using
both of them at the same time.

If any of the properties is missing, that most likely needs
to be fixed in the .ui file.

As a side note, something similar to the previous algorithm to
determine the grid position might be useful to implement
QtInstanceWidget::get_grid_left_attach etc. at some point.

Change-Id: I22e34275e3974f42ca3ee43fa6cb66624e917443
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/175661
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>