tdf#130857 qt weld: Set actual positions for layouts inside grid
Extend QtBuilder::applyGridPackingProperties to not only
handle widgets (QWidget), but also layouts (QLayout), so
that the proper position inside the grid is also set for
these.

This will be used e.g. by the "Help" -> "About LibreOfficeDev"
dialog.
(When adding "cui/ui/aboutdialog.ui" to the list of .ui files
in QtInstanceBuilder::IsUIFileSupported, those UI elements
that show up are now in the correct positions, but other things
unrelated to this particular commit are still missing.)

Change-Id: Ic7c342e0307526f29e1fa149acd25a27373069ff
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/175662
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>