tdf#130857 qt weld: Implement QtInstanceImage::set_image
Extract a helper to convert a BitmapEx to a QPixmap
from the existing helper function `loadQPixmapIcon`
and add a new helper that converts a
css::uno::Reference<css::graphic::XGraphic>& to
a QPixmap reusing most of the code.

Use that to implement the QtInstanceImage::set_image
variant with that parameter type.

With this in place, opening the "Help" -> "About LibreOfficeDev"
dialog in a WIP branch adding "cui/ui/aboutdialog.ui" to the list
of .ui files in QtInstanceBuilder::IsUIFileSupported now opens
the dialog with brand image and LibreOffice Community logo present,
but other aspects in that dialog still need attention before
enabling this by default.

Change-Id: Ide28c0dc126be511ec8e34e57f4f1df7e70297ca
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/175666
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>