tdf#130857 qt weld: Evaluate + set grid item row/col span
When processing packing properties for grid items,
also take the "width" and "height" properties into account,
which specify the column span and row span, respectively.

Pass these as params as well when re-adding the item
to the grid, not just the row/col index.

With this in place, when opening the "Help" -> "About LibreOfficeDev"
dialog in a WIP branch adding "cui/ui/aboutdialog.ui" to the list
of .ui files in QtInstanceBuilder::IsUIFileSupported, the
"imAbout" image on the left hand side of the dialog is now no longer
restricted to the top left cell in the grid, but uses more space
as expected.

Change-Id: Ia2edb147d7576d863d597c01d4d310e8798a9ecc
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/175668
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>