tdf#130857 qt weld: Make link buttons focusable + handle activation
Set the text interaction flags [1] for the QLabel used for
hyperlinks so that the links can be highlighted and activated
using mouse and keyboard.

Connect to the QLabel::linkActivated signal that gets emitted
when a link gets activated, and call signal_activate_link
in response. If that returns false, open the URL using
QDesktopServices::openUrl instead.

Another way to only support opening the URL using
QDesktopServices::openUrl automatically would be
to set the `openExternalLinks` property [2], but
IIUC, the custom activation handler should have
priority.

With this in place, when opening the "Help" -> "About LibreOfficeDev"
dialog in a WIP branch adding "cui/ui/aboutdialog.ui" to the list
of .ui files in QtInstanceBuilder::IsUIFileSupported, the tab
order now includes the hyperlinks in the dialog, and pressing
Enter while one is focused, or clicking on one of them opens them
in the browser when using the qt6 VCL plugin.

[1] https://doc.qt.io/qt-6/qlabel.html#textInteractionFlags-prop
[2] https://doc.qt.io/qt-6/qlabel.html#openExternalLinks-prop

Change-Id: I4e3cbca6d01a652a1038b9750ee9049e1c7491be
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/175670
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins