tdf#130857 qt weld: Don't expect packing props on implicit grid
When the "left-attach" or "top-attach" packing
properties are not set in QtBuilder::applyGridPackingProperties,
just skip processing, but don't trigger an assert.

While these packing properties should be set for every
child of a "GtkGrid" object in .ui files, there's also
the special case of QMessageBox, which uses a QGridLayout
for its layout whereas the .ui file for a "GtkMessageDialog"
uses a "GtkBox" instead.
(See the special handling for "GtkBox" in QtBuilder::makeObject,
where the QMessageBox's existing layout is used instead
of setting another one.)

Therefore, gracefully handle the case where the
properties are not set. Without this, the "Save Document?"
dialog triggered the assert since

    commit 3e28b4f0c97d011cf4222941f019ce05fe6b0313
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Fri Oct 25 21:21:25 2024 +0200

        tdf#130857 qt weld: Simplify QtBuilder::applyGridPackingProperties

Change-Id: I03bc9799c142563ffbf0fa3eadb58bc63e6e23f7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/175722
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>