Fix unresponsive mouse moved events in Writer on macOS
On macOS, when Writer is in a background task such as initial layout
and spellchecking, mouse moved events appear to be unresponsive due
to the fact that AquaSalInstance::AnyInput() returns false when flags
include VclInputFlags::MOUSE even though there are native mouse moved
events waiting to be dispatched.

In contrast, on Windows WinSalInstance::AnyInput() will return true
so make macOS behave the same as Windows and return true if there
are any queued native mouse moved events.

Change-Id: I4b3cf66f2fb0357c0eebaf869da0bdcdbb8f8516
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/175620
Reviewed-by: Adolfo Jayme Barrientos <fitojb@ubuntu.com>
Tested-by: Jenkins
Reviewed-by: Patrick Luby <guibomacdev@gmail.com>