win a11y: Drop "Enable AT Tool support" from options UI
Drop the "Support assistive technology tools (program restart
required)" checkbox from the "Accessibility" options page.

It was only shown on Windows.

While having this in the UI seems to suggest that this
is a setting that users explicitly toggle on and off,
the setting is automatically switched on when an AT is active
anyway, see `ImplHandleGetObject` in
vcl/win/window/salframe.cxx.

Therefore, only switching explicitly off after using
an AT, but no longer planning to do so any more is
the only case I can think of where changing this
option via the UI might have been a use case.

The plan is to rework the handling in an upcoming
commit instead: to enable AT support when an AT
is active, i.e. in `ImplHandleGetObject`, and not
persist that setting any more across LO runs.

Related change for the Windows registry:

    commit 77e1c0bfa2e643f017b829cd775d08c5a40d9b6b
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Wed Oct 23 12:17:02 2024 +0100

        win a11y: No longer read/write registry key

Change-Id: I32624b830d39d08510e4731edd06fd7a77642c50
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/175787
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>